generate-maphub-seeddata script upload
Files:
1. generate-maphub-seeddata.rb script
2. samplescreenApril8.png
3. yamlseeddata.png

Description:

Uploaded the script which runs. I also attached a screenshot showing it
running and the sample yaml file with the given output as required.

The next step is to test it on the server which I couldn't do given
that I was working from home this week.

Stress Tested: 170 image files (which is the random sample that I had
downloaded from the server to my local machine for development)